reorganized email login code
